Refactor modules and add package metadata
