+++ archivecli/__init__.py
@@ -1,0 +1,5 @@
+"""archivecli package metadata."""
+
+__all__ = ["__version__"]
+__version__ = "0.1.0"
+
